**Please provide a description of this PR:**
Follow-up of https://github.com/istio/istio/pull/41610
Creating a new PR for the alternative optimizations I made to `MostSpecificHostMatch`.
This inverts the idea behind the method, instead of iterating all wildcard hosts to find one that matches, we construct all the possible wildcards that would match on the given needle and attempt to find those in the wildcard map.

@hzxuzhonghu @howardjohn since you reviewed the last PR

This **greatly** reduces the cpu usage of the method, but does incur in more allocations:
```
name                                                      old time/op    new time/op    delta
MostSpecificHostMatch/10ExactNoMatch-8                      15.0ns ± 3%    10.6ns ± 1%  -29.04%  (p=0.000 n=10+9)
MostSpecificHostMatch/50ExactNoMatch-8                      14.7ns ± 0%    10.7ns ± 1%  -27.54%  (p=0.000 n=9+10)
MostSpecificHostMatch/100ExactNoMatch-8                     19.2ns ± 0%    15.2ns ± 1%  -21.00%  (p=0.000 n=9+10)
MostSpecificHostMatch/1000ExactNoMatch-8                    14.7ns ± 1%    10.6ns ± 0%  -27.89%  (p=0.000 n=10+10)
MostSpecificHostMatch/5000ExactNoMatch-8                    14.8ns ± 2%    10.7ns ± 1%  -28.08%  (p=0.000 n=9+9)
MostSpecificHostMatch/10ExactMatch-8                        9.35ns ± 1%    9.40ns ± 2%   +0.47%  (p=0.038 n=8+9)
MostSpecificHostMatch/50ExactMatch-8                        15.0ns ± 1%    11.9ns ± 2%  -20.71%  (p=0.000 n=9+10)
MostSpecificHostMatch/100ExactMatch-8                       10.0ns ± 1%    12.0ns ± 1%  +19.49%  (p=0.000 n=10+9)
MostSpecificHostMatch/1000ExactMatch-8                      9.35ns ± 1%   10.02ns ± 1%   +7.21%  (p=0.000 n=9+10)
MostSpecificHostMatch/5000ExactMatch-8                      9.78ns ± 1%   11.60ns ± 1%  +18.62%  (p=0.000 n=9+9)
MostSpecificHostMatch/10DestRuleWildcardNoMatch-8            136ns ± 1%     121ns ± 4%  -11.15%  (p=0.000 n=9+10)
MostSpecificHostMatch/50DestRuleWildcardNoMatch-8            465ns ± 1%     120ns ± 1%  -74.07%  (p=0.000 n=9+8)
MostSpecificHostMatch/100DestRuleWildcardNoMatch-8          1.13µs ± 1%    0.13µs ± 1%  -88.77%  (p=0.000 n=9+8)
MostSpecificHostMatch/1000DestRuleWildcardNoMatch-8         12.9µs ± 1%     0.1µs ± 3%  -99.09%  (p=0.000 n=10+9)
MostSpecificHostMatch/5000DestRuleWildcardNoMatch-8         54.7µs ± 0%     0.1µs ± 1%  -99.78%  (p=0.000 n=10+9)
MostSpecificHostMatch/10DestRuleWildcardMatch-8              149ns ± 0%      39ns ± 2%  -73.93%  (p=0.000 n=10+9)
MostSpecificHostMatch/50DestRuleWildcardMatch-8              633ns ± 1%      46ns ± 4%  -92.77%  (p=0.000 n=10+10)
MostSpecificHostMatch/100DestRuleWildcardMatch-8            1.17µs ± 1%    0.04µs ± 1%  -96.43%  (p=0.000 n=9+10)
MostSpecificHostMatch/1000DestRuleWildcardMatch-8           12.9µs ± 0%     0.0µs ± 2%  -99.70%  (p=0.000 n=9+10)
MostSpecificHostMatch/5000DestRuleWildcardMatch-8           67.2µs ± 1%     0.0µs ± 2%  -99.94%  (p=0.000 n=9+9)
MostSpecificHostMatch/10NeedleWildcardNoMatch-8              144ns ± 2%     124ns ± 1%  -13.28%  (p=0.000 n=10+10)
MostSpecificHostMatch/50NeedleWildcardNoMatch-8              594ns ± 1%     142ns ± 1%  -76.09%  (p=0.000 n=9+9)
MostSpecificHostMatch/100NeedleWildcardNoMatch-8            1.17µs ± 0%    0.13µs ± 1%  -88.60%  (p=0.000 n=9+10)
MostSpecificHostMatch/1000NeedleWildcardNoMatch-8           12.8µs ± 1%     0.1µs ± 1%  -99.00%  (p=0.000 n=10+9)
MostSpecificHostMatch/5000NeedleWildcardNoMatch-8           55.2µs ± 0%     0.1µs ± 1%  -99.78%  (p=0.000 n=10+9)
MostSpecificHostMatch/10NeedleWildcardMatch-8                155ns ± 1%      44ns ± 0%  -71.46%  (p=0.000 n=9+8)
MostSpecificHostMatch/50NeedleWildcardMatch-8                617ns ± 0%      47ns ± 1%  -92.39%  (p=0.000 n=9+9)
MostSpecificHostMatch/100NeedleWildcardMatch-8              1.16µs ± 0%    0.05µs ± 1%  -95.99%  (p=0.000 n=10+10)
MostSpecificHostMatch/1000NeedleWildcardMatch-8             12.8µs ± 0%     0.0µs ± 1%  -99.64%  (p=0.000 n=8+9)
MostSpecificHostMatch/5000NeedleWildcardMatch-8             67.7µs ± 1%     0.0µs ± 0%  -99.93%  (p=0.000 n=9+8)
MostSpecificHostMatchMixed/10DestRuleWildcardNoMatch-8       143ns ± 1%     126ns ± 1%  -11.36%  (p=0.000 n=10+10)
MostSpecificHostMatchMixed/50DestRuleWildcardNoMatch-8       480ns ± 0%     134ns ± 1%  -72.12%  (p=0.000 n=10+9)
MostSpecificHostMatchMixed/100DestRuleWildcardNoMatch-8     1.17µs ± 0%    0.14µs ± 1%  -88.05%  (p=0.000 n=9+9)
MostSpecificHostMatchMixed/1000DestRuleWildcardNoMatch-8    12.8µs ± 0%     0.1µs ± 2%  -99.03%  (p=0.000 n=9+9)
MostSpecificHostMatchMixed/5000DestRuleWildcardNoMatch-8    54.9µs ± 0%     0.1µs ± 1%  -99.77%  (p=0.000 n=10+9)
MostSpecificHostMatchMixed/10DestRuleWildcardMatch-8         157ns ± 2%      44ns ± 1%  -71.65%  (p=0.000 n=10+9)
MostSpecificHostMatchMixed/50DestRuleWildcardMatch-8         657ns ± 1%      49ns ± 0%  -92.52%  (p=0.000 n=9+10)
MostSpecificHostMatchMixed/100DestRuleWildcardMatch-8       1.16µs ± 1%    0.05µs ± 1%  -96.06%  (p=0.000 n=10+9)
MostSpecificHostMatchMixed/1000DestRuleWildcardMatch-8      12.9µs ± 0%     0.0µs ± 1%  -99.62%  (p=0.000 n=10+10)
MostSpecificHostMatchMixed/5000DestRuleWildcardMatch-8      67.3µs ± 1%     0.1µs ± 5%  -99.93%  (p=0.000 n=9+9)
MostSpecificHostMatchMixed/10NeedleWildcardNoMatch-8         135ns ± 1%     124ns ± 2%   -7.83%  (p=0.000 n=9+10)
MostSpecificHostMatchMixed/50NeedleWildcardNoMatch-8         586ns ± 0%     132ns ± 1%  -77.41%  (p=0.000 n=10+9)
MostSpecificHostMatchMixed/100NeedleWildcardNoMatch-8       1.18µs ± 0%    0.13µs ± 1%  -89.08%  (p=0.000 n=10+9)
MostSpecificHostMatchMixed/1000NeedleWildcardNoMatch-8      13.0µs ± 0%     0.1µs ± 2%  -98.99%  (p=0.000 n=9+9)
MostSpecificHostMatchMixed/5000NeedleWildcardNoMatch-8      54.9µs ± 0%     0.1µs ± 0%  -99.77%  (p=0.000 n=9+8)
MostSpecificHostMatchMixed/10NeedleWildcardMatch-8           148ns ± 1%      45ns ± 1%  -69.98%  (p=0.000 n=10+9)
MostSpecificHostMatchMixed/50NeedleWildcardMatch-8           636ns ± 1%      51ns ± 0%  -91.92%  (p=0.000 n=10+9)
MostSpecificHostMatchMixed/100NeedleWildcardMatch-8         1.17µs ± 0%    0.05µs ± 1%  -95.84%  (p=0.000 n=9+9)
MostSpecificHostMatchMixed/1000NeedleWildcardMatch-8        12.9µs ± 1%     0.0µs ± 0%  -99.65%  (p=0.000 n=10+8)
MostSpecificHostMatchMixed/5000NeedleWildcardMatch-8        67.6µs ± 0%     0.0µs ± 0%  -99.93%  (p=0.000 n=10+10)
MostSpecificHostMatchMultiMatch/DestRuleWildcard-8          86.9ns ± 0%    35.9ns ± 1%  -58.76%  (p=0.000 n=10+9)
MostSpecificHostMatchMultiMatch/NeedleWildcard-8            91.4ns ± 0%    39.5ns ± 2%  -56.78%  (p=0.000 n=9+9)

name                                                      old alloc/op   new alloc/op   delta
MostSpecificHostMatch/10ExactNoMatch-8                       0.00B          0.00B          ~     (all equal)
MostSpecificHostMatch/50ExactNoMatch-8                       0.00B          0.00B          ~     (all equal)
MostSpecificHostMatch/100ExactNoMatch-8                      0.00B          0.00B          ~     (all equal)
MostSpecificHostMatch/1000ExactNoMatch-8                     0.00B          0.00B          ~     (all equal)
MostSpecificHostMatch/5000ExactNoMatch-8                     0.00B          0.00B          ~     (all equal)
MostSpecificHostMatch/10ExactMatch-8                         0.00B          0.00B          ~     (all equal)
MostSpecificHostMatch/50ExactMatch-8                         0.00B          0.00B          ~     (all equal)
MostSpecificHostMatch/100ExactMatch-8                        0.00B          0.00B          ~     (all equal)
MostSpecificHostMatch/1000ExactMatch-8                       0.00B          0.00B          ~     (all equal)
MostSpecificHostMatch/5000ExactMatch-8                       0.00B          0.00B          ~     (all equal)
MostSpecificHostMatch/10DestRuleWildcardNoMatch-8            0.00B         32.00B ± 0%    +Inf%  (p=0.000 n=10+10)
MostSpecificHostMatch/50DestRuleWildcardNoMatch-8            0.00B         32.00B ± 0%    +Inf%  (p=0.000 n=10+10)
MostSpecificHostMatch/100DestRuleWildcardNoMatch-8           0.00B         32.00B ± 0%    +Inf%  (p=0.000 n=10+10)
MostSpecificHostMatch/1000DestRuleWildcardNoMatch-8          0.00B         32.00B ± 0%    +Inf%  (p=0.000 n=10+10)
MostSpecificHostMatch/5000DestRuleWildcardNoMatch-8          0.00B         32.00B ± 0%    +Inf%  (p=0.000 n=10+10)
MostSpecificHostMatch/10DestRuleWildcardMatch-8              0.00B         16.00B ± 0%    +Inf%  (p=0.000 n=10+10)
MostSpecificHostMatch/50DestRuleWildcardMatch-8              0.00B         16.00B ± 0%    +Inf%  (p=0.000 n=10+10)
MostSpecificHostMatch/100DestRuleWildcardMatch-8             0.00B         24.00B ± 0%    +Inf%  (p=0.000 n=10+10)
MostSpecificHostMatch/1000DestRuleWildcardMatch-8            0.00B         24.00B ± 0%    +Inf%  (p=0.000 n=10+10)
MostSpecificHostMatch/5000DestRuleWildcardMatch-8            0.00B         24.00B ± 0%    +Inf%  (p=0.000 n=10+10)
MostSpecificHostMatch/10NeedleWildcardNoMatch-8              0.00B         32.00B ± 0%    +Inf%  (p=0.000 n=10+10)
MostSpecificHostMatch/50NeedleWildcardNoMatch-8              0.00B         32.00B ± 0%    +Inf%  (p=0.000 n=10+10)
MostSpecificHostMatch/100NeedleWildcardNoMatch-8             0.00B         32.00B ± 0%    +Inf%  (p=0.000 n=10+10)
MostSpecificHostMatch/1000NeedleWildcardNoMatch-8            0.00B         32.00B ± 0%    +Inf%  (p=0.000 n=10+10)
MostSpecificHostMatch/5000NeedleWildcardNoMatch-8            0.00B         32.00B ± 0%    +Inf%  (p=0.000 n=10+10)
MostSpecificHostMatch/10NeedleWildcardMatch-8                0.00B         16.00B ± 0%    +Inf%  (p=0.000 n=10+10)
MostSpecificHostMatch/50NeedleWildcardMatch-8                0.00B         16.00B ± 0%    +Inf%  (p=0.000 n=10+10)
MostSpecificHostMatch/100NeedleWildcardMatch-8               0.00B         24.00B ± 0%    +Inf%  (p=0.000 n=10+10)
MostSpecificHostMatch/1000NeedleWildcardMatch-8              0.00B         24.00B ± 0%    +Inf%  (p=0.000 n=10+10)
MostSpecificHostMatch/5000NeedleWildcardMatch-8              0.00B         24.00B ± 0%    +Inf%  (p=0.000 n=10+10)
MostSpecificHostMatchMixed/10DestRuleWildcardNoMatch-8       0.00B         32.00B ± 0%    +Inf%  (p=0.000 n=10+10)
MostSpecificHostMatchMixed/50DestRuleWildcardNoMatch-8       0.00B         32.00B ± 0%    +Inf%  (p=0.000 n=10+10)
MostSpecificHostMatchMixed/100DestRuleWildcardNoMatch-8      0.00B         32.00B ± 0%    +Inf%  (p=0.000 n=10+10)
MostSpecificHostMatchMixed/1000DestRuleWildcardNoMatch-8     0.00B         32.00B ± 0%    +Inf%  (p=0.000 n=10+10)
MostSpecificHostMatchMixed/5000DestRuleWildcardNoMatch-8     0.00B         32.00B ± 0%    +Inf%  (p=0.000 n=10+10)
MostSpecificHostMatchMixed/10DestRuleWildcardMatch-8         0.00B         16.00B ± 0%    +Inf%  (p=0.000 n=10+10)
MostSpecificHostMatchMixed/50DestRuleWildcardMatch-8         0.00B         16.00B ± 0%    +Inf%  (p=0.000 n=10+10)
MostSpecificHostMatchMixed/100DestRuleWildcardMatch-8        0.00B         24.00B ± 0%    +Inf%  (p=0.000 n=10+10)
MostSpecificHostMatchMixed/1000DestRuleWildcardMatch-8       0.00B         24.00B ± 0%    +Inf%  (p=0.000 n=10+10)
MostSpecificHostMatchMixed/5000DestRuleWildcardMatch-8       0.00B         24.00B ± 0%    +Inf%  (p=0.000 n=10+10)
MostSpecificHostMatchMixed/10NeedleWildcardNoMatch-8         0.00B         32.00B ± 0%    +Inf%  (p=0.000 n=10+10)
MostSpecificHostMatchMixed/50NeedleWildcardNoMatch-8         0.00B         32.00B ± 0%    +Inf%  (p=0.000 n=10+10)
MostSpecificHostMatchMixed/100NeedleWildcardNoMatch-8        0.00B         32.00B ± 0%    +Inf%  (p=0.000 n=10+10)
MostSpecificHostMatchMixed/1000NeedleWildcardNoMatch-8       0.00B         32.00B ± 0%    +Inf%  (p=0.000 n=10+10)
MostSpecificHostMatchMixed/5000NeedleWildcardNoMatch-8       0.00B         32.00B ± 0%    +Inf%  (p=0.000 n=10+10)
MostSpecificHostMatchMixed/10NeedleWildcardMatch-8           0.00B         16.00B ± 0%    +Inf%  (p=0.000 n=10+10)
MostSpecificHostMatchMixed/50NeedleWildcardMatch-8           0.00B         16.00B ± 0%    +Inf%  (p=0.000 n=10+10)
MostSpecificHostMatchMixed/100NeedleWildcardMatch-8          0.00B         24.00B ± 0%    +Inf%  (p=0.000 n=10+10)
MostSpecificHostMatchMixed/1000NeedleWildcardMatch-8         0.00B         24.00B ± 0%    +Inf%  (p=0.000 n=10+10)
MostSpecificHostMatchMixed/5000NeedleWildcardMatch-8         0.00B         24.00B ± 0%    +Inf%  (p=0.000 n=10+10)
MostSpecificHostMatchMultiMatch/DestRuleWildcard-8           0.00B         24.00B ± 0%    +Inf%  (p=0.000 n=10+10)
MostSpecificHostMatchMultiMatch/NeedleWildcard-8             0.00B         24.00B ± 0%    +Inf%  (p=0.000 n=10+10)

name                                                      old allocs/op  new allocs/op  delta
MostSpecificHostMatch/10ExactNoMatch-8                        0.00           0.00          ~     (all equal)
MostSpecificHostMatch/50ExactNoMatch-8                        0.00           0.00          ~     (all equal)
MostSpecificHostMatch/100ExactNoMatch-8                       0.00           0.00          ~     (all equal)
MostSpecificHostMatch/1000ExactNoMatch-8                      0.00           0.00          ~     (all equal)
MostSpecificHostMatch/5000ExactNoMatch-8                      0.00           0.00          ~     (all equal)
MostSpecificHostMatch/10ExactMatch-8                          0.00           0.00          ~     (all equal)
MostSpecificHostMatch/50ExactMatch-8                          0.00           0.00          ~     (all equal)
MostSpecificHostMatch/100ExactMatch-8                         0.00           0.00          ~     (all equal)
MostSpecificHostMatch/1000ExactMatch-8                        0.00           0.00          ~     (all equal)
MostSpecificHostMatch/5000ExactMatch-8                        0.00           0.00          ~     (all equal)
MostSpecificHostMatch/10DestRuleWildcardNoMatch-8             0.00           3.00 ± 0%    +Inf%  (p=0.000 n=10+10)
MostSpecificHostMatch/50DestRuleWildcardNoMatch-8             0.00           3.00 ± 0%    +Inf%  (p=0.000 n=10+10)
MostSpecificHostMatch/100DestRuleWildcardNoMatch-8            0.00           3.00 ± 0%    +Inf%  (p=0.000 n=10+10)
MostSpecificHostMatch/1000DestRuleWildcardNoMatch-8           0.00           3.00 ± 0%    +Inf%  (p=0.000 n=10+10)
MostSpecificHostMatch/5000DestRuleWildcardNoMatch-8           0.00           3.00 ± 0%    +Inf%  (p=0.000 n=10+10)
MostSpecificHostMatch/10DestRuleWildcardMatch-8               0.00           1.00 ± 0%    +Inf%  (p=0.000 n=10+10)
MostSpecificHostMatch/50DestRuleWildcardMatch-8               0.00           1.00 ± 0%    +Inf%  (p=0.000 n=10+10)
MostSpecificHostMatch/100DestRuleWildcardMatch-8              0.00           1.00 ± 0%    +Inf%  (p=0.000 n=10+10)
MostSpecificHostMatch/1000DestRuleWildcardMatch-8             0.00           1.00 ± 0%    +Inf%  (p=0.000 n=10+10)
MostSpecificHostMatch/5000DestRuleWildcardMatch-8             0.00           1.00 ± 0%    +Inf%  (p=0.000 n=10+10)
MostSpecificHostMatch/10NeedleWildcardNoMatch-8               0.00           3.00 ± 0%    +Inf%  (p=0.000 n=10+10)
MostSpecificHostMatch/50NeedleWildcardNoMatch-8               0.00           3.00 ± 0%    +Inf%  (p=0.000 n=10+10)
MostSpecificHostMatch/100NeedleWildcardNoMatch-8              0.00           3.00 ± 0%    +Inf%  (p=0.000 n=10+10)
MostSpecificHostMatch/1000NeedleWildcardNoMatch-8             0.00           3.00 ± 0%    +Inf%  (p=0.000 n=10+10)
MostSpecificHostMatch/5000NeedleWildcardNoMatch-8             0.00           3.00 ± 0%    +Inf%  (p=0.000 n=10+10)
MostSpecificHostMatch/10NeedleWildcardMatch-8                 0.00           1.00 ± 0%    +Inf%  (p=0.000 n=10+10)
MostSpecificHostMatch/50NeedleWildcardMatch-8                 0.00           1.00 ± 0%    +Inf%  (p=0.000 n=10+10)
MostSpecificHostMatch/100NeedleWildcardMatch-8                0.00           1.00 ± 0%    +Inf%  (p=0.000 n=10+10)
MostSpecificHostMatch/1000NeedleWildcardMatch-8               0.00           1.00 ± 0%    +Inf%  (p=0.000 n=10+10)
MostSpecificHostMatch/5000NeedleWildcardMatch-8               0.00           1.00 ± 0%    +Inf%  (p=0.000 n=10+10)
MostSpecificHostMatchMixed/10DestRuleWildcardNoMatch-8        0.00           3.00 ± 0%    +Inf%  (p=0.000 n=10+10)
MostSpecificHostMatchMixed/50DestRuleWildcardNoMatch-8        0.00           3.00 ± 0%    +Inf%  (p=0.000 n=10+10)
MostSpecificHostMatchMixed/100DestRuleWildcardNoMatch-8       0.00           3.00 ± 0%    +Inf%  (p=0.000 n=10+10)
MostSpecificHostMatchMixed/1000DestRuleWildcardNoMatch-8      0.00           3.00 ± 0%    +Inf%  (p=0.000 n=10+10)
MostSpecificHostMatchMixed/5000DestRuleWildcardNoMatch-8      0.00           3.00 ± 0%    +Inf%  (p=0.000 n=10+10)
MostSpecificHostMatchMixed/10DestRuleWildcardMatch-8          0.00           1.00 ± 0%    +Inf%  (p=0.000 n=10+10)
MostSpecificHostMatchMixed/50DestRuleWildcardMatch-8          0.00           1.00 ± 0%    +Inf%  (p=0.000 n=10+10)
MostSpecificHostMatchMixed/100DestRuleWildcardMatch-8         0.00           1.00 ± 0%    +Inf%  (p=0.000 n=10+10)
MostSpecificHostMatchMixed/1000DestRuleWildcardMatch-8        0.00           1.00 ± 0%    +Inf%  (p=0.000 n=10+10)
MostSpecificHostMatchMixed/5000DestRuleWildcardMatch-8        0.00           1.00 ± 0%    +Inf%  (p=0.000 n=10+10)
MostSpecificHostMatchMixed/10NeedleWildcardNoMatch-8          0.00           3.00 ± 0%    +Inf%  (p=0.000 n=10+10)
MostSpecificHostMatchMixed/50NeedleWildcardNoMatch-8          0.00           3.00 ± 0%    +Inf%  (p=0.000 n=10+10)
MostSpecificHostMatchMixed/100NeedleWildcardNoMatch-8         0.00           3.00 ± 0%    +Inf%  (p=0.000 n=10+10)
MostSpecificHostMatchMixed/1000NeedleWildcardNoMatch-8        0.00           3.00 ± 0%    +Inf%  (p=0.000 n=10+10)
MostSpecificHostMatchMixed/5000NeedleWildcardNoMatch-8        0.00           3.00 ± 0%    +Inf%  (p=0.000 n=10+10)
MostSpecificHostMatchMixed/10NeedleWildcardMatch-8            0.00           1.00 ± 0%    +Inf%  (p=0.000 n=10+10)
MostSpecificHostMatchMixed/50NeedleWildcardMatch-8            0.00           1.00 ± 0%    +Inf%  (p=0.000 n=10+10)
MostSpecificHostMatchMixed/100NeedleWildcardMatch-8           0.00           1.00 ± 0%    +Inf%  (p=0.000 n=10+10)
MostSpecificHostMatchMixed/1000NeedleWildcardMatch-8          0.00           1.00 ± 0%    +Inf%  (p=0.000 n=10+10)
MostSpecificHostMatchMixed/5000NeedleWildcardMatch-8          0.00           1.00 ± 0%    +Inf%  (p=0.000 n=10+10)
MostSpecificHostMatchMultiMatch/DestRuleWildcard-8            0.00           1.00 ± 0%    +Inf%  (p=0.000 n=10+10)
MostSpecificHostMatchMultiMatch/NeedleWildcard-8              0.00           1.00 ± 0%    +Inf%  (p=0.000 n=10+10)
```