@howardjohn 2022-11-01
Not sure this code path is actually used enough that this matters enough to make it this complex (does it?), but looks like all allocs are about doing `"*."+host`. Can we just store with without `*.` prefix?

Something like:
```
diff --git a/pilot/pkg/model/cluster_local.go b/pilot/pkg/model/cluster_local.go
index 7dd6eb20a6..5c041a724b 100644
--- a/pilot/pkg/model/cluster_local.go
+++ b/pilot/pkg/model/cluster_local.go
@@ -105,7 +105,11 @@ func (c *clusterLocalProvider) onMeshUpdated(e *Environment) {
 			for _, h := range serviceSettings.GetHosts() {
 				hostname := host.Name(h)
 				if hostname.IsWildCarded() {
-					hosts.wildcard[hostname] = struct{}{}
+					if hostname == "*" {
+						hosts.wildcard[""] = struct{}{}
+					} else {
+						hosts.wildcard[hostname[2:]] = struct{}{}
+					}
 				} else {
 					hosts.specific[hostname] = struct{}{}
 				}
diff --git a/pilot/pkg/model/config.go b/pilot/pkg/model/config.go
index ef332d5492..7e96bc37cb 100644
--- a/pilot/pkg/model/config.go
+++ b/pilot/pkg/model/config.go
@@ -283,7 +283,7 @@ func MostSpecificHostMatch[V any](needle host.Name, specific map[host.Name]V, wi
 
 	if needle.IsWildCarded() {
 		// exact match first
-		if v, ok := wildcard[needle]; ok {
+		if v, ok := wildcard[needle[2:]]; ok {
 			return needle, v, true
 		}
 
@@ -321,14 +321,14 @@ func mostSpecificHostWildcardMatch[V any](needle string, wildcard map[host.Name]
 		needle = needle[i+1:]
 
 		// atempt to find wildcard match for this part
-		wild := host.Name("*." + needle)
+		wild := host.Name(needle)
 		if v, ok := wildcard[wild]; ok {
 			return wild, v, true
 		}
 	}
 
 	// attempt to find global wildcard match
-	wild := host.Name("*")
+	wild := host.Name("")
 	if v, ok := wildcard[wild]; ok {
 		return wild, v, true
 	}
```

(note: that doesn't pass tests)