@hzxuzhonghu 2022-11-02
Yes, looks this is not the hot code path, and it increases memory allocations